Rename file to match class + remove dead code
